Refactor to avoid circular imports
