+++ tests/test_split_matrix.py
@@ -7,7 +7,7 @@ import scipy.sparse as sps
 
 import tabmat as tm
 from tabmat import from_pandas
-from tabmat.constructor import _split_sparse_and_dense_parts
+from tabmat.constructor_util import _split_sparse_and_dense_parts
 from tabmat.dense_matrix import DenseMatrix
 from tabmat.ext.sparse import csr_dense_sandwich
 from tabmat.split_matrix import SplitMatrix
